refactor: uploading without leaving the current page
